# Change Dynamic header underlying data type from byte[] to string