--- a/src/libraries/Common/src/System/Net/Http/Http2/Hpack/HPackDecoder.cs
+++ b/src/libraries/Common/src/System/Net/Http/Http2/Hpack/HPackDecoder.cs
@@ -93,10 +93,9 @@ namespace System.Net.Http.HPack
         private byte[] _headerValueOctets;
 
         private State _state = State.Ready;
-        private byte[] _headerName;
+        private string _headerName;
         private int _stringIndex;
         private int _stringLength;
-        private int _headerNameLength;
         private int _headerValueLength;
         private bool _index;
         private bool _huffman;
@@ -374,14 +373,13 @@ namespace System.Net.Http.HPack
         {
             OnString(nextState: State.Ready);
 
-            var headerNameSpan = new Span<byte>(_headerName, 0, _headerNameLength);
-            var headerValueSpan = new Span<byte>(_headerValueOctets, 0, _headerValueLength);
+            var headerValue = WebHeaderEncoding.GetString(_headerValueOctets, 0, _headerValueLength);
 
-            handler?.OnHeader(headerNameSpan, headerValueSpan);
+            handler?.OnHeader(_headerName, headerValue);
 
             if (_index)
             {
-                _dynamicTable.Insert(headerNameSpan, headerValueSpan);
+                _dynamicTable.Insert(_headerName, headerValue);
             }
         }
 
@@ -405,7 +403,6 @@ namespace System.Net.Http.HPack
         {
             HeaderField header = GetHeader(index);
             _headerName = header.Name;
-            _headerNameLength = header.Name.Length;
             _state = State.HeaderValueLength;
         }
 
@@ -450,8 +447,9 @@ namespace System.Net.Http.HPack
             {
                 if (_state == State.HeaderName)
                 {
-                    _headerNameLength = Decode(ref _headerNameOctets);
-                    _headerName = _headerNameOctets;
+                    var headerNameLength = Decode(ref _headerNameOctets);
+                    _headerName = WebHeaderEncoding.GetString(_headerNameOctets, 0, headerNameLength);
+
                 }
                 else
                 {
@@ -466,13 +464,13 @@ namespace System.Net.Http.HPack
             _state = nextState;
         }
 
-        private HeaderField GetHeader(int index)
+        private ref HeaderField GetHeader(int index)
         {
             try
             {
-                return index <= H2StaticTable.Count
-                    ? H2StaticTable.Get(index - 1)
-                    : _dynamicTable[index - H2StaticTable.Count - 1];
+                return ref index <= H2StaticTable.Count
+                    ? ref H2StaticTable.Get(index - 1)
+                    : ref _dynamicTable[index - H2StaticTable.Count - 1];
             }
             catch (IndexOutOfRangeException)
             {