--- a/src/libraries/System.Net.Http/src/System/Net/Http/Headers/HeaderDescriptor.cs
+++ b/src/libraries/System.Net.Http/src/System/Net/Http/Headers/HeaderDescriptor.cs
@@ -4,6 +4,7 @@
 
 using System.Buffers;
 using System.Diagnostics;
+using System.Text;
 using System.Text.Unicode;
 
 namespace System.Net.Http.Headers
@@ -169,6 +170,57 @@ namespace System.Net.Http.Headers
 
             decoded = null;
             return false;
+        }
+
+        public string GetHeaderValue(string headerValue)
+        {
+            if (headerValue.Length == 0)
+            {
+                return string.Empty;
+            }
+
+            // If it's a known header value, use the known value instead of allocating a new string.
+            if (_knownHeader != null)
+            {
+                if (_knownHeader.KnownValues != null)
+                {
+                    string[] knownValues = _knownHeader.KnownValues;
+                    for (int i = 0; i < knownValues.Length; i++)
+                    {
+                        if (knownValues[i].Equals(headerValue, StringComparison.OrdinalIgnoreCase))
+                        {
+                            return knownValues[i];
+                        }
+                    }
+                }
+
+                if (_knownHeader == KnownHeaders.Location)
+                {
+                    Debug.Assert(true, "This overload is used only by HTTP2 so this slow path shouldn't never happen.");
+                    // Normally Location should be in ISO-8859-1 but occasionally some servers respond with UTF-8.
+                    if (TryDecodeUtf8(headerValue, out string decoded))
+                    {
+                        return decoded;
+                    }
+                }
+            }
+
+            return headerValue;
+        }
+
+        // The string is considered Latin 1 encoding
+        private static bool TryDecodeUtf8(string input, out string decoded)
+        {
+            try
+            {
+                decoded = Encoding.UTF8.GetString(Encoding.GetEncoding("latin1").GetBytes(input));
+                return true;
+            }
+            catch (ArgumentException)
+            {
+                decoded = null;
+                return false;
+            }
         }
 
         // QPack Static Table