# Perf results:

## Dynamic table
Insertion into dynamic table:

### Before:

| Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|------:|------:|------:|----------:|
| Insert | **1.459 us** | 0.0700 us | 0.1974 us |     - |     - |     - |      **88 B** |

### After:

| Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| Insert | **234.7 ns** | 12.73 ns | 37.35 ns |     - |     - |     - |      **48 B** |

## HeaderField constructor

There was removed the main allocation issue

### Before:
| Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
|    Ctr | **336.0 ns** | 6.66 ns | 7.13 ns | 0.1121 |     - |     - |     **704 B** |
### After:
| Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|------:|------:|------:|----------:|
|    Ctr | **21.86 ns** | 0.107 ns | 0.101 ns |     - |     - |     - |         **-** |

##HeaderDescriptor

I added new string overload for *GetHeaderValue* method as the original with byte[] is used by HTTP2 and there was a performance penalty to convert the value to string before.

As you can see there is a significant performance hit for Location headers, however, this path is not used in HTTP2. I would like to discuss this in the PR.

### Before:
|                             Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
|       GetHeaderValue_Bytes_Unknown |  **34.96 ns** | 0.355 ns | 0.315 ns | 0.0114 |     - |     - |      **72 B** |
|         GetHeaderValue_Bytes_Known |  **46.35 ns** | 0.296 ns | 0.247 ns | 0.0114 |     - |     - |      **72 B** |
|      GetHeaderValue_Bytes_Location |  **87.41 ns** | 0.995 ns | 0.931 ns | 0.0114 |     - |     - |      **!!** *72 B* |
| GetHeaderValue_Bytes_Location_utf8 | **109.67 ns** | 2.222 ns | 5.366 ns | 0.0114 |     - |     - |  **!!**    *72 B* |
### After:
|                              Method |       Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------ |-----------:|----------:|----------:|-------:|------:|------:|----------:|
|       GetHeaderValue_String_Unknown |   **3.719 ns** | 0.0983 ns | 0.1009 ns |      - |     - |     - |         **-** |
|         GetHeaderValue_String_Known |   **4.268 ns** | 0.0332 ns | 0.0294 ns |      - |     - |     - |         **-** |
|      GetHeaderValue_String_Location | **!! 208.229 ns** | 3.1741 ns | 2.8138 ns | 0.0203 |     - |     - |     **!!**    *128 B* |
| GetHeaderValue_String_Location_utf8 | **!! 242.598 ns** | 2.5850 ns | 2.2916 ns | 0.0200 |     - |     - |     **!!**    *128 B* |

## HpackDecoder

It is visible that the decoder doesn't allocate in case of dynamic/static headers at all

### Before:

|                                 Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| DecodesIndexedHeaderField_DynamicTable | **81.19 ns** | 1.149 ns | 1.075 ns | 0.0153 |     - |     - |      **96 B **|
|  DecodesIndexedHeaderField_StaticTable | **76.35 ns** | 1.746 ns | 2.331 ns | 0.0114 |     - |     - |      **72 B** |

### After:

|                                 Method |     Mean |    Error |   StdDev |   Median | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------------- |---------:|---------:|---------:|---------:|------:|------:|------:|----------:|
| DecodesIndexedHeaderField_DynamicTable | **14.61 ns** | 0.501 ns | 0.419 ns | 14.45 ns |     - |     - |     - |         **-** |
|  DecodesIndexedHeaderField_StaticTable | **13.42 ns** | 0.313 ns | 0.808 ns | 13.10 ns |     - |     - |     - |         **-** |

## Http client

The main improvement is in simple get with an empty body (*Get*) as Kestrel sends value from the static table. Another requests with custom headers are equal as this change is not related to encoding (*GetWithHeaders*) and the Kestrel don't send dynamic headers in the response (*GetWithHeaders_Echo*)

### Before:

|              Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
|                 Get | **449.6 us** | 10.03 us | 28.13 us | 0.4883 |     - |     - |   **4.58 KB** |
|      GetWithHeaders | 467.2 us | 10.36 us | 30.55 us |      - |     - |     - |   5.11 KB |
| GetWithHeaders_Echo | 469.3 us |  9.25 us | 13.84 us | 0.9766 |     - |     - |   6.53 KB |

### After:

|              Method |     Mean |   Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|---------:|-------:|------:|------:|----------:|
|                 Get | **344.2 us** | 7.69 us |  9.44 us | 0.4883 |     - |     - |   **3.76 KB** |
|      GetWithHeaders | 431.6 us | 8.55 us |  9.50 us | 0.4883 |     - |     - |   5.12 KB |
| GetWithHeaders_Echo | 468.8 us | 9.19 us | 12.58 us | 0.9766 |     - |     - |   6.53 KB |

Fixes #31861