@Tratcher 2020-02-06
Did you run this by anyone in AspNet or try out the changes there? This is going to be very disruptive, and it might prevent us from delivering our custom decoder feature https://github.com/dotnet/aspnetcore/issues/17400.