@aik-jahoda 2020-03-13
Closing as it requires more discussion with ASP net how to attack this issue (it is in conflict with dotnet/aspnetcore#17400).